requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dataclasses-json==0.5.14
dnspython==2.4.2
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Mail==0.9.1
frozenlist==1.4.0
google==3.0.0
google-api-core==2.11.1
google-auth==2.22.0
google-cloud==0.34.0
google-cloud-aiplatform==1.32.0
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-cloud-resource-manager==1.10.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
google-search-results==2.4.2
googleapis-common-protos==1.60.0
grpc-google-iam-v1==0.12.6
grpcio==1.57.0
grpcio-status==1.57.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.277
langsmith==0.0.29
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.25.2
openai==0.27.10
packaging==23.1
proto-plus==1.22.3
protobuf==4.24.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
Shapely==1.8.5.post1
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.20
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.3.7
yarl==1.9.2
